added dark level + spikes prefab
created spike via blender and imported to unity.
spikes move up after a delay when player steps on them.
if the spikes hit the player then they a random direction  impluse force will be applied on the player.

added dark level where the only light is a spot light that follows player.

it seems like there are a lot of changes because i reorganized the folder of the project. (moved all scripts into scripts folder, etc..)