character can now move on left and right walls
movingTileCtrl is for level 2. it's pretty much a tile that moves front and back. (still work in progress)

the movement on the walls is working pefrectly but the rotation that the player makes to reach the wall still needs some work. (you need to move the character a bit once you get on the wall in order to get his rotation right)

for now pressing Ctrl changes gravity to UP
pressing ALT toggles between right and left wall.

did it like this for now because we aren't entirely sure about how we want to do this in the future.